2025_10_16_How_to_Build_an_Agent/pyproject.toml
[project]
name = "agentic-ai"
version = "0.1.0"
description = "An example project explaining how to build agents"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.70.0",
    "beautifulsoup4>=4.14.2",
    "fastmcp>=2.12.4",
    "h2ogpte==1.6.42",
    "jinja2>=3.1.6",
    "jupyterlab>=4.4.9",
    "kaleido>=0.2.1",
    "openai>=2.3.0",
    "pandas>=2.1.0",
    "pillow>=10.0.0",
    "plotly>=5.18.0",
    "python-dotenv>=1.1.1",
    "tabulate>=0.9.0",
]
